Feat: add functionality to insultForm
